Crate serial_port

Support for basic serial port access, including initialization, transmit, and receive.

Resources used for this implementation:

https://en.wikibooks.org/wiki/Serial_Programming/8250_UART_Programming

https://tldp.org/HOWTO/Modem-HOWTO-4.html

https://wiki.osdev.org/Serial_Ports

Notes

Some serial port drivers use special cases for transmitting some byte values, specifically 0x08 and 0x7F, which are ASCII “backspace” and “delete”, respectively. They do so by writing them as three distinct values (with proper busy waiting in between):

0x08

0x20 (an ascii space character)

0x08 again.

This isn’t necessarily a bad idea, as it “clears out” whatever character was there before, presumably to prevent rendering/display issues for a deleted character. But, this isn’t required, and I personally believe it should be handled by a higher layer, such as a shell or TTY program. We don’t do anything like that here, in case a user of this crate wants to send binary data across the serial port, rather than “smartly-interpreted” ASCII characters.

Structs

SerialPort
A serial port and its various data and control registers.

Constants

COM1_BASE_PORT

COM2_BASE_PORT

COM3_BASE_PORT

COM4_BASE_PORT