Function serial_port::init_serial_port

pub fn init_serial_port(
    serial_port_address: SerialPortAddress, 
    serial_port: SerialPortBasic
) -> &'static Arc<MutexIrqSafe<SerialPort>>

Initializes the SerialPort specified by the given SerialPortAddress.

This function also registers the interrupt handler for this serial port such that it can receive data using interrupts instead of busy-waiting or polling.

If the given serial port has already been initialized, this does nothing and simply returns a reference to the already-initialized serial port.