Struct serial_port::SerialPort

pub struct SerialPort { /* private fields */ }

A serial port abstraction with support for interrupt-based data receival.

Implementations

impl SerialPort

pub fn new(serial_port: SerialPortBasic) -> SerialPort

Initialize this serial port by giving it ownership and control of the given basic serial_port.

pub fn register_interrupt_handler(
serial_port: Arc<MutexIrqSafe<SerialPort>>,
interrupt_number: u8,
interrupt_handler: HandlerFunc
) -> Result<(), &'static str>

Register the interrupt handler for this serial port and spawn a deferrent interrupt task to handle its data receival.

pub fn set_data_sender(&mut self, sender: Sender<DataChunk>) -> Result<(), ()>

Tells this SerialPort to push received data bytes onto the given sender channel.

If a sender already exists for this serial port, the existing sender is not replaced and an error is returned.